Pull vsd_configt out into its own file
