+++ src/analyses/Makefile
@@ -60,6 +60,7 @@ SRC = ai.cpp \
       variable-sensitivity/value_set_pointer_abstract_object.cpp \
       variable-sensitivity/value_set_array_abstract_object.cpp \
       variable-sensitivity/three_way_merge_abstract_interpreter.cpp \
+      variable-sensitivity/variable_sensitivity_configuration.cpp \
       # Empty last line
 
 INCLUDES= -I ..
